optimize: new date formater for format date str.
